Automatic Irrigating System
This is the simple code in Arduino  which helps to irrigating system to make automatic. For your project you can create your own code using this sample code. this is the sample code of our project. I am not showing the circuit diagram for this project. You can draw yourself as you need. If any queries  then comment I will give you the solution.